==> 03-cleaning.Rmd <==
# Data transformation
In our data transformation script, we perform the following steps:

1. Since all data are `.csv` files, we load them using the `read.csv()` function in R. 

2. We drop the unnecessary variables in the bike-sharing data: `Duration`, `Start Station`, `Rnd Station`, and `Bike number`. 

3. We choose the variable `Start Date` to represent the occurrence time of the bike riding. It is separated into four date and time variables: `Year`, `Month`, `Day`, and `Hour`. 

4. For the temperature data, we change the format of the variable `DATE` to the standard practice `YYYY-MM-DD`.

5. We merge the two data sets on the index date variable by using `left_join()` function in the `dplyr` package. 

6. We add the variable `workdays`, which is true if the day is not a holiday or weekend and false otherwise.

7. Since the size of the data frame is too large with every row representing a single trip, we change the granularity to hour by aggregation over the total count of bike riding trips in each hour of a day. 

The derived data set is condensed to 16,889 rows with 13 variables:

* Year - Year of the occurrence of the bike riding, 2019
* Month - Month of the occurrence of the bike riding
* Day - Day of the occurrence of the bike riding
* Hour - Hour of the occurrence of the bike riding
* PRCP - The amount of pricipitation
* SNOW - The amount of snow
* SNWD - The depth of snow
* TAVG - The average temperature in Fahrenheit
* TMAX - The maximum temperature in Fahrenheit
* TMIN - The minimum temperature in Fahrenheit
* workdays - Indicating if it is a workday
* Member.Type – Indicates whether user was a "registered" member or a "casual" rider
* num - Number of bike riding


```{r message=FALSE, warning=FALSE}
library(dplyr)
library(purrr)
library(prob)
library(ggplot2)
library(ggridges)
library(tidyverse)
library(lubridate)
```

```{r message=FALSE, warning=FALSE, results = 'hide'}
# create data folder
if (!file.exists("data")) {
  dir.create(file.path(getwd(), "data"))
}

# download bike sharing data
address = "https://s3.amazonaws.com/capitalbikeshare-data/201907-capitalbikeshare-tripdata.zip"
download.file(address, file.path("data", basename(address)))
unzip(file.path("data", basename(address)), "201907-capitalbikeshare-tripdata", exdir=file.path("data"))
file.rename(file.path("data", "201907-capitalbikeshare-tripdata"), file.path("data", "201907-capitalbikeshare-tripdata.csv"))
bike <<- read.csv("data/201907-capitalbikeshare-tripdata.csv")
for (m in 1:12) {
  if (m == 7) {
    next
  }
  address = sprintf("https://s3.amazonaws.com/capitalbikeshare-data/2019%02d-capitalbikeshare-tripdata.zip",m)
  download.file(address, file.path("data", basename(address)))
  f = unzip(file.path("data", basename(address)), gsub(".zip", ".csv", basename(address)), exdir=file.path("data"))
  bike = rbind(bike, read.csv(f))
}


# drop unnecessary variables
bike = bike %>% select(c(1, 2, 3, 9))

# create variables regarding date and time
t_start = strptime(as.character(bike$Start.date), "%Y-%m-%d %H:%M:%S")
bike$start_date = as.character(date(t_start))
bike$Year = year(t_start)
bike$Month = month(t_start, label = TRUE)
bike$Day = day(t_start)
bike$Hour = hour(t_start)
bike = bike %>% select(!c("Start.date", "End.date"))

# read temperature data
# change date format
temp = read.csv("temp_data/temp.csv")
temp$DATE = format(strptime(as.character(temp$DATE), "%m/%d/%Y"), "%Y-%m-%d")

# merge two data frames on date
bike = left_join(bike, temp, by = c("start_date" = "DATE"))

# add the variable: day of week
bike$weekday = weekdays(as.Date(bike$start_date, "%Y-%m-%d"))

# holidays of 2011 and 2012
holidays = c("2019-01-01","2019-01-21", "2019-02-18", "2019-04-13", "2019-05-27", "2019-07-04", "2019-09-02",
             "2019-10-14", "2019-11-11", "2019-11-28", "2019-12-25")

# add a variable to check if the bike sharing happened on a workday
bike$holidays = bike$start_date %in% holidays
bike$workdays = !(bike$weekday %in% c("Saturday", "Sunday") | bike$holidays)
bike$workdays = recode_factor(as.factor(bike$workdays), "TRUE" = "Workdays", "FALSE" = "Non-workdays")
bike = bike %>% select(!c("weekday", "holidays", "start_date", "Duration"))

bike = bike %>% 
  group_by(Year, Month, Day, Hour, PRCP, SNOW, SNWD, TAVG, TMAX, TMIN, workdays, Member.type) %>%
  count(name="num")
write.csv(bike, "data/bike.csv")
```
