==> 04-missing.Rmd <==
# Missing values
After running the code `data[!complete.cases(data), ]`, we list all rows with missing values and find out that there are no such rows. Therefore, no imputation is necessary.

```{r}
df = read.csv("data/bike.csv")
df[!complete.cases(df), ]
```
