==> 05-results.Rmd <==
# Results

```{r}
# Load packages for results
library(dplyr)
library(purrr)
library(prob)
library(ggplot2)
library(ggridges)
library(tidyverse)
library(lubridate)
```


## Part 1
The granularity of the main data is at the hourly level. However, for some of the analysis we will also want to compute daily statistics. In particular, in this part we will be analyzing the daily number of registered and unregistered users.

```{r}
# Data for part 1
folder_path = "data/bike_data"
csv_path = "data/201901-capitalbikeshare-tripdata.csv"
df_original = read.csv(csv_path)
for (m in 2:12) {
  csv_path = sprintf("data/2019%02d-capitalbikeshare-tripdata.csv", m)
  df_original = rbind(df_original, read.csv(csv_path))
}
df = df_original %>% select(c(1, 2, 3, 9))
t_start = strptime(as.character(df$Start.date), "%Y-%m-%d %H:%M:%S")
df$date = as.character(date(t_start))
df$year = year(t_start)
df$month = month(t_start, label=TRUE)
df$day = day(t_start)
df$hour = hour(t_start)
df = df %>% select(!c("Start.date", "End.date"))
df$weekday = weekdays(as.Date(df$date, "%Y-%m-%d"), abbreviate=TRUE)
holidays = c("2019-01-01","2019-01-21", "2019-02-18", "2019-04-13", "2019-05-27", "2019-07-04", "2019-09-02",
             "2019-10-14", "2019-11-11", "2019-11-28", "2019-12-25")
df$holiday = df$date %in% holidays
df$workingday = !(df$weekday %in% c("Sat", "Sun") | df$holiday)
df$holiday = recode_factor(as.factor(df$holiday), "TRUE" = "yes", "FALSE" = "no")
df$workingday = recode_factor(as.factor(df$workingday), "TRUE" = "yes", "FALSE" = "no")
```


### Distributions of the daily counts of casual and member riders
Let's begin by comparing the distribution of the daily counts of casual and member riders. The following plot overlays the distribution of the daily counts of `casual` and `member` user with the granularity of the records being daily counts.

```{r}
plot_multi_histogram <- function(df, feature, label_column) {
    plt <- ggplot(df, aes(x=eval(parse(text=feature)), fill=eval(parse(text=label_column)))) +
        geom_histogram(aes(y = ..density..), alpha=0.7, position="identity", color="black", binwidth=500) +
        geom_density(alpha=0.7) +
        labs(x="Rider count", y="Density", caption="Data source: Capital Bikeshare")
    plt + guides(fill=guide_legend(title="Member type")) +
        ggtitle("Distribution Comparison of Casual vs Registered Riders") +
      scale_y_continuous(labels = scales::comma)
}
daily_counts = df %>% group_by(Member.type, date) %>% count()
plot_multi_histogram(daily_counts, 'n', 'Member.type')
```

The distribution for the `casual` riders is right skewed with the mode at around 1,500, a tail extending to over 5,000 (no counts more than 7,500); the one with the `member` ones is left skewed with the mode at around 11,000. The spread of the `casual` distribution is more concentrated on the range from 0 to 4,000 than that of the `member` one, which has a spread of much wider span from 0 to 15,000. Both distribution do not have significant outliers.

### Relationship between the daily counts of casual and member riders
The density plots do not show us how the daily counts for `casual` and `member` riders vary together. The following scatter plot helps us to investigate the relationship between `casual` and `member` counts. We will also draw a linear linear regression line for both groups. The points in the scatter plot are colored according to whether or not the day is working day. Since there are many points in the scatter plot, we make them small to help with solving the overplotting issue.

```{r}
daily_counts <- df %>% 
  group_by(Member.type, date, workingday) %>% 
  count()
daily_counts <- daily_counts %>% pivot_wider(names_from=Member.type, values_from=n)
ggplot(daily_counts, aes(x=Casual, y=Member, color=workingday)) +
  geom_point() +
  geom_smooth(method=lm) +
    ggtitle("Comparison of Casual vs Registered Riders on Working and Nonworking Days") +
    labs(caption="Data source: Capital Bikeshare")
```

The scatter plot roughly exhibits a linear relationship between `casual` and `registered` riders on the weekend. Moreover, it makes clear the separation between the work days and non-work days. However, the overplotting makes it difficult to see the density of the joint counts. To address this issue, let's try visualizing the data with another technique, the kernel density plot.


```{r}
ggplot(daily_counts, aes(x=Casual, y=Member, group=workingday)) + 
  stat_density_2d(aes(alpha=..level.., fill=workingday), geom="polygon") +
  labs(caption="Data source: Capital Bikeshare") +
  ggtitle("KDE Plot Comparison of Registered vs Casual Riders")
```

The kernel density plot suggests that for work days there are significantly more `member` riders than `casual` ones but for nonwork days the difference is not as large. It's obviously easier to see the relationship on this plot since the density centers look more clear and therefore the overplotting issue is addressed.


## Part 2

### Number of riders every hour in Jan 2019
```{r}
bike = read.csv("data/bike.csv")
date = paste(bike$Year, bike$Month, bike$Day, sep="-")
time = paste(bike$Hour, "00", "00", sep = ":")
bike$time = as.POSIXct(paste(date, time), format="%Y-%b-%d %H:%M:%S")
bike %>% 
  mutate(Member.type = factor(Member.type)) %>%
  mutate(Member.type = fct_rev(Member.type)) %>%
  dplyr::filter(Month == "Jan") %>%
  ggplot(aes(x=time, y=num, color=Member.type)) +
  geom_line() +
  facet_wrap(.~(Day > 15), ncol=1, scales="free_x") +
  ggtitle("Number of riders every hour in Jan 2019") +
  labs(x="Time", y="Number", caption="Data source: Capital Bikeshare") +
  theme(strip.text.x = element_blank(), plot.title = element_text(lineheight=1, face="bold", hjust = 0.5)) +
  scale_color_manual(values = c("skyblue3", "lightpink3"))
bike = bike %>% select(!time)
```

If we look at the graph of the number of riders every hour in Jan 2019, we notice that there is a repeating pattern that the number of riders keep fluctuating every day. Let's take a closer look at the pattern of number of bike riding every day. We want to investigate the average number of bike riding each hour in 2019.

### Daily pattern of bike riding
```{r}
timeline = c("0:00","1:00","2:00","3:00","4:00","5:00","6:00","7:00","8:00",
             "9:00","10:00","11:00","12:00","13:00","14:00","15:00","16:00",
             "17:00","18:00","19:00","20:00","21:00","22:00","23:00")
bike %>% 
  group_by(Hour) %>%
  summarise(num=mean(num)) %>%
  mutate(Time = factor(paste(Hour,"00",sep = ":"), levels = timeline)) %>%
  mutate(Time = fct_rev(Time)) %>%
  ggplot(aes(x=Time, y=num)) +
    geom_bar(stat="identity", fill="steelblue") +
    ylab("Number of riders") +
    labs(caption="Data source: Capital Bikeshare") +
    ggtitle("Average number of riders each hour") +
    theme(plot.title = element_text(lineheight=1, face="bold", hjust = 0.5)) +
    coord_flip()
```

According to the graph above, it can be observed that the number of riders peaks during rush hours each day. Moreover, the number of riders gradually decreases and achieves its lowest point at 3:00 AM. 

### Difference between patterns of riders
```{r}
bike %>% 
  group_by(Hour, Member.type) %>% 
  summarise(num=mean(num)) %>%
  mutate(Time = factor(sprintf("%s:00",Hour))) %>%
  mutate(Time = fct_rev(Time)) %>%
  ggplot(aes(x=Time, y=num)) + 
    geom_bar(stat="identity", fill="steelblue") +
    ggtitle("Average number of two types of riders each hour") +
    labs(caption="Data source: Capital Bikeshare") +
    theme_grey(13) +
    theme(plot.title = element_text(lineheight=1, face="bold", hjust = 0.5)) +
    coord_flip() +
    facet_wrap(~Member.type)
```

When comparing the number of casual and member riders, we notice that the former is much lower than the latter. More importantly, there are different patterns of bike riding for these two types of riders. For members, as in the last graph, the number of riders increases rapidly during rush hours. It can be inferred that many member riders use bike sharing for daily commute. On the other hand, for casual riders, most bike sharing happen during the daytime and there is no rapid increase in rush hours.


## Part 3

In this part of analysis, let's find out how the weather condition influences the riding behaviors. The data granularity is at the daily level. 

```{r}
df<-read.csv("https://raw.githubusercontent.com/r3qin/bikesharing/main/data/bike.csv")
total_sum<-df %>% group_by(Year,Month,Day) %>% summarize(number=sum(num)) 
day<-df %>% group_by(Year,Month,Day) %>% slice_head(n=1)
by_day<-day %>% left_join(total_sum,by=c('Year','Month','Day'),keep=FALSE) %>% subset(select=-c(X,Hour,num,Member.type));
```

### Average number of trips per day in different temperatures

Firstly, we want to visualize the average number of trips per day in different temperatures.

```{r}
library(ggplot2)
temp_count<- by_day %>% group_by(TAVG) %>% summarize(num=mean(number))
ggplot(temp_count,aes(TAVG,num))+
  geom_col()+
  labs(title="Relationship between temperature and daily trip counts",y="average trips per day",x="average temperature (Fahrenheit)")+
  scale_x_continuous(breaks=c(20,30,40,50,60,70,80,90,100))+
  theme_grey(14)+
  labs(caption="Data source: Capital Bikeshare") +
  theme(plot.title = element_text(face="bold"))
```

The graph below clearly shows that the number of trips ramps up quickly from 0 to 60 degrees, but the relationship between bike rental and temperature becomes weaker in temperatures above 60 degrees. The mode is at around 55 degree, meaning that there is a distinctly high number of bike rentals in that temperature. After some further investigation, we found out that a potential explanation for the unusual spike is that the National cherry blossom festival in DC was at the day for the mode temperature.

### Average daily trips counts with different precipitation

Next, let's visualize the relationship between rainy days and the number of trips. 

```{r}
prcp_count<- by_day %>% group_by(PRCP,workdays) %>% summarize(num=mean(number));
ggplot(prcp_count,aes(x=PRCP,y=num,color=workdays))+
  geom_point(alpha=0.5)+
  geom_smooth(se = FALSE)+
  facet_wrap(vars(workdays),nrow=1)+
  labs(title="Average daily trip counts with different precipitation",
       y="Average trips per day",x="PRCP(inches)", 
       caption="Data source: Capital Bikeshare")+
    theme_gray(14)+
  theme(plot.title = element_text(face="bold"))
```

It's counter-intuitive to see that the number of trips is not necessarily less in the rainy days. In fact, we can see a few outliers at the days with extremely heavy rains have a pretty high number of trips. Also, it's interesting to see that the number of trips on workdays are less affected by rains than on weekends. A reasonable guess is that on weekdays people cannot avoid renting bikes for commuting no matter if it is rainy or not.

### Average daily trips counts by snow

Then, we looked into the impact of snow on the number of trips. 

```{r}
library(forcats)
by_snowfall<-by_day %>% mutate(ifsnow=if_else(SNOW>0,"Snowy","Not snowy") ) 
ggplot(by_snowfall,aes(x=as_factor(ifsnow),y=number,fill=ifsnow))+
  geom_boxplot(varwidth=TRUE)+
labs(title="Comparision of daily trip counts by snow",y="Number of trips",fill="Weather")+
  scale_fill_manual(values=c("#E69F00", "#56B4E9"))+
  theme_grey(14)+
  labs(caption="Data source: Capital Bikeshare")+
  theme(plot.title = element_text(face="bold"),
        axis.title.x = element_blank())
```

```{r}
ggplot(by_day[by_day$SNWD!=0,],aes(x=SNWD,y=number))+
  geom_point(aes(color=workdays))+
  labs(title="Average daily trip counts with different snowdepth",y="Number of trips",x="Snow Depth (inches)")+
  theme_grey(14)+
  labs(caption="Data source: Capital Bikeshare")+
  theme(plot.title = element_text(face="bold"))
```

The impact of snow is very obvious as the average number of trips drops drastically on snowy days. Additionally, though the data points for the snowy days are scarce, we can roughly observe that with the increasing severeness of snow, the number of trips decreases proportionally. And during work days the snow depth has a smaller effect on the number of trips than during non-work days.
