==> 06-interactive.Rmd <==
# Interactive component

```{r echo=FALSE, message=FALSE, warning=FALSE}
# Data preparatrion for interactive graph
library(dplyr)
library(purrr)
library(prob)
library(ggplot2)
library(ggridges)
library(tidyverse)
library(lubridate)
bike = read.csv("data/bike.csv")
Q_1 = c("Jan", "Feb", "Mar")
Q_2 = c("Apr", "May", "Jun")
Q_3 = c("Jul", "Aug", "Sep")
Q_4 = c("Oct", "Nov", "Dec")
hour_inter_1 = bike %>% 
  dplyr::filter(Month %in% Q_1) %>%
  group_by(Hour, Member.type) %>%
  summarise(num=mean(num)) %>%
  pivot_wider(names_from = Member.type, values_from = num) %>%
  mutate_if(is.numeric, round) %>%
  select(Casual, Member) %>%
  rename(Casual_1 = Casual, Member_1 = Member)
hour_inter_2 = bike %>% 
  dplyr::filter(Month %in% Q_2) %>%
  group_by(Hour, Member.type) %>%
  summarise(num=mean(num)) %>%
  pivot_wider(names_from = Member.type, values_from = num) %>%
  mutate_if(is.numeric, round) %>%
  select(Casual, Member) %>%
  rename(Casual_2 = Casual, Member_2 = Member)
hour_inter_3 = bike %>% 
  dplyr::filter(Month %in% Q_3) %>%
  group_by(Hour, Member.type) %>%
  summarise(num=mean(num)) %>%
  pivot_wider(names_from = Member.type, values_from = num) %>%
  mutate_if(is.numeric, round) %>%
  select(Casual, Member) %>%
  rename(Casual_3 = Casual, Member_3 = Member)
hour_inter_4 = bike %>% 
  dplyr::filter(Month %in% Q_4) %>%
  group_by(Hour, Member.type) %>%
  summarise(num=mean(num)) %>%
  pivot_wider(names_from = Member.type, values_from = num) %>%
  mutate_if(is.numeric, round) %>%
  select(Casual, Member) %>%
  rename(Casual_4 = Casual, Member_4 = Member)
hour_inter_5 = bike %>% 
  group_by(Hour, Member.type) %>%
  summarise(num=mean(num)) %>%
  pivot_wider(names_from = Member.type, values_from = num) %>%
  mutate_if(is.numeric, round) %>%
  select(Casual, Member) %>%
  rename(Casual_5 = Casual, Member_5 = Member)
inter = cbind(hour_inter_1, hour_inter_2, hour_inter_3, hour_inter_4, hour_inter_5)
write.csv(inter, "data/inter.csv", row.names = FALSE)
```



## Daily pattern of bike riding

Hovering the mouse on any bar in the following graph, we can see the hour of a day and the average number of riders during that hour. By selecting among different quarters and the whole year of 2019, we can check the average number of bike riding each hour over this selected period of time. 

<script src="https://d3js.org/d3.v6.js"></script>
<svg width="1000" height="700" id="part2_svg"></svg>
<script src="part2.js"></script>

Comparing graphs of different quarters, we observe that people tend to use the bike sharing more during summer time than during winter. Also, the average number of bike riding of `member` riders peaks at 8:00 AM and 17:00 PM in all four quarters and that of `casual` riders increases in the daytime. In short, the riding patterns for `member` and `casual` riders are very constant across the entire year. 

## Weather effect

Since most of the precipitation in the rainy days in DC is less than 1 inches, points are clustered towards the left of the static graph of precipitation. It is thus difficult to tell whether the association of trip counts with rain is the same across different levels of rains. To further explore the effect of rain on the daily riders' counts, we created an interactive scatter plot where we can zoom in and out of the x axis using the mouse and touch pad. The interaction allows us to look into the impact of specific range of precipitation on the daily riders' counts. 

<svg width="960" height="500" id="part3_svg"></svg>
<script src="part3.js"></script>

In fact, we found that precipitation has a diminishing marginal effect. Specifically, the pattern of change of daily trip counts from 0 (no rain) to 0.1 inches of rain is more obvious than that from 0.5 to 0.6 inches.
