==> 07-conclusion.Rmd <==
# Conclusion

Due to the scarce occurrences of the unfavorable weather conditions in DC within a year, there exists only a small number of data points for snowy days, weakening our conclusion for snow depth's negative association with bike rental counts. Additionally, since most days in DC in 2019 have precipitation between 0 and 1 inches per day, the visualization for precipitation is limited to a tight scope. In order to have a more convincing study of weather's impact on ridership, we need to use data from cities with more snowy days or a wider range of precipitation to solve the data deficiency and imbalance issue respectively. Enriching the trip counts in snowy days and diversifying the range of precipitation will both help produce clearer patterns in the plots.

Moreover, because of our choice of data granularities, we discarded other important elements such as duration, start and end locations from the original data source. In the future, we could try to gain more insights into the spatial and temporal information by takings these elements into consideration. A list of topics will be intriguing to explore, such as the weather's effect on the trip duration, the most common geographical riding path for both types of users and etc. 

Overall, there are multiple facets from which we are able to study bike-sharing system, not only in DC, but also across the globe. The city dynamics are definitely different among different areas. In conclusion, our biggest takeaway from this project is that as we investigate a topic through the lens of data, we should take diverse attempts with an open and cautious mindset.
